requirements.txt
Pillow
icoextract
pefile
